# api/openapi.yaml
openapi: 3.0.3
info:
  title: Simplus API
  version: 0.1.0
  description: Locale-neutral control-plane contract for Simplus.
servers:
  - url: /
paths:
...
  /api/v1/messages:
    get:
      operationId: listMessages
      summary: List the most recent durable SMS records
      tags: [messages]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
        - name: lineId
          in: query
          description: Optional stable Line identity; remoteAddress is required when lineId is present
          schema: { type: string, pattern: "^line_[A-Za-z0-9_-]{22}$" }
        - name: remoteAddress
          in: query
          description: Optional exact conversation peer; without lineId, returns the peer history across Lines
          schema: { type: string, pattern: "^(?:\\+?[0-9]{3,20}|[A-Za-z][A-Za-z0-9 ._-]{0,19})$" }
      responses:
        "200":
          description: SMS records in newest-first order by their first successful local persistence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMSMessageListResponse"
        "401":
          description: No live administrator session is present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "400":
          description: Pagination cursor, limit, or conversation filter is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The instance is not ready for business APIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "421":
          description: The HTTP authority is not a permitted loopback host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Message history could not be read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Messaging is not configured for this runtime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "504":
          description: The request exceeded the bounded control-plane deadline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: sendMessage
      summary: Persist and send one SMS on a selected line
      tags: [messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendSMSRequest"
      responses:
        "200":
          description: An idempotent replay of an existing operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMSMessage"
        "201":
          description: The SMS operation was created and reached its current durable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMSMessage"
        "400":
          description: The operation id, line, destination, or body is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: No live administrator session is present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: CSRF validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: The selected line is not present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The line is unavailable, the instance is not ready, or the operation id conflicts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "421":
          description: The HTTP authority is not a permitted loopback host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "422":
          description: The selected line does not support SMS
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: The durable SMS state could not be saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: The SMS transport or inventory is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "504":
          description: The request exceeded the bounded control-plane deadline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v1/message-conversations:
    get:
      operationId: listMessageConversations
      summary: List durable SMS conversations by exact remote address
      tags: [messages]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: SMS conversations in newest-first order by each latest message's first successful local persistence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SMSConversationListResponse"
        "400":
          description: Pagination cursor or limit is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: No live administrator session is present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The instance is not ready for business APIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "421":
          description: The HTTP authority is not a permitted trusted-LAN host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Message conversations could not be read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Messaging is not configured for this runtime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "504":
          description: The request exceeded the bounded control-plane deadline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /api/v1/message-conversations/read-state:
    put:
      operationId: markMessageConversationRead
      summary: Mark one visible SMS conversation snapshot as read
      tags: [messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkSMSConversationReadRequest"
      responses:
        "204":
          description: The snapshot boundary is read or was already read
        "400":
          description: The remote address or opaque read boundary is invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: No live administrator session is present
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "403":
          description: CSRF validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: The boundary message no longer exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The instance is not ready for business APIs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "421":
          description: The HTTP authority is not a permitted trusted-LAN host
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: The read state could not be persisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Messaging is not configured for this runtime
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "504":
          description: The request exceeded the bounded control-plane deadline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /api/v1/calls:
    get:
      operationId: listCalls
      summary: List Simulator call records
      tags: [calls]
      parameters:
        - $ref: "#/components/parameters/PageLimit"
        - $ref: "#/components/parameters/PageCursor"
      responses:
        "200":
          description: Recent call records
          content: { application/json: { schema: { $ref: "#/components/schemas/CallListResponse" } } }
        "400": { description: Pagination cursor or limit is invalid, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: No administrator session, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { description: Instance not ready, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "500": { description: Calls unavailable, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "421": { description: The HTTP authority is not a permitted trusted-LAN host, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "503": { description: Calls are not configured for this runtime, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "504": { description: The request exceeded the bounded control-plane deadline, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
...
  /api/v1/calls/dial:
    post:
      operationId: dialCall
      summary: Dial in the Simulator call state machine
      tags: [calls]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/CallStartRequest" } } }
      responses:
        "200": { description: Idempotent replay, content: { application/json: { schema: { $ref: "#/components/schemas/Call" } } } }
        "201": { description: Call created, content: { application/json: { schema: { $ref: "#/components/schemas/Call" } } } }
        "400": { description: Invalid or unsafe number, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "401": { description: No administrator session, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "409": { description: Line unavailable or replay conflict, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
        "500": { description: Call persistence failed, content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } } }
...
components:
  parameters:
    PageLimit:
      name: limit
      in: query
      description: Maximum records returned; defaults to 20
      schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
    PageCursor:
      name: cursor
      in: query
      description: Versioned opaque cursor from the previous page
      schema: { type: string, minLength: 1, maxLength: 256, pattern: "^[A-Za-z0-9_-]+$" }
  schemas:
...
    SMSMessageListResponse:
      type: object
      additionalProperties: false
      required: [messages, totalCount, capacity, nearCapacity]
      properties:
        messages:
          type: array
          maxItems: 50
          items:
            $ref: "#/components/schemas/SMSMessage"
        totalCount:
          type: integer
          format: int64
          minimum: 0
        capacity:
          type: integer
          format: int64
          minimum: 1
        nearCapacity:
          type: boolean
        nextCursor:
          type: string
          minLength: 1
          maxLength: 256
          pattern: "^[A-Za-z0-9_-]+$"
        readThroughToken:
          type: string
          minLength: 1
          maxLength: 256
          pattern: "^[A-Za-z0-9_-]+$"
...
    ApiError:
      type: object
      additionalProperties: false
      required: [code, retryable]
      properties:
        code:
          type: string
          pattern: "^[A-Z0-9_]+$"
        retryable:
          type: boolean
        reference:
          type: string